# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NOVA Ride API",
    "version": "1.0.0",
    "description": "جرد مسارات مُولّد آليًا من المتحكّمات. البادئة /api؛ كل المسارات متاحة أيضًا تحت /api/v1. لا يتضمّن هذا الجرد مخطّطات الطلب/الاستجابة (تُثرى لاحقًا)."
  },
  "servers": [
    {
      "url": "{baseUrl}",
      "variables": {
        "baseUrl": {
          "default": "http://localhost:4000"
        }
      }
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/driver/me/trips/:id": {
      "get": {
        "tags": [
          "drivers"
        ],
        "operationId": "drivers_trip",
        "summary": "GET /api/driver/me/trips/:id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/growth/experiments": {
      "get": {
        "tags": [
          "growth"
        ],
        "operationId": "growth_listExperiments",
        "summary": "GET /api/growth/experiments",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "growth"
        ],
        "operationId": "growth_createExperiment",
        "summary": "POST /api/growth/experiments",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/growth/experiments/:key/assignment": {
      "get": {
        "tags": [
          "growth"
        ],
        "operationId": "growth_assign",
        "summary": "GET /api/growth/experiments/:key/assignment",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/growth/incentives": {
      "get": {
        "tags": [
          "growth"
        ],
        "operationId": "growth_listIncentives",
        "summary": "GET /api/growth/incentives",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "growth"
        ],
        "operationId": "growth_createIncentive",
        "summary": "POST /api/growth/incentives",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/growth/incentives/:id/progress": {
      "post": {
        "tags": [
          "growth"
        ],
        "operationId": "growth_recordProgress",
        "summary": "POST /api/growth/incentives/:id/progress",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/message-templates": {
      "get": {
        "tags": [
          "message-templates"
        ],
        "operationId": "message-templates_list",
        "summary": "GET /api/message-templates",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: notifications.send"
      },
      "post": {
        "tags": [
          "message-templates"
        ],
        "operationId": "message-templates_create",
        "summary": "POST /api/message-templates",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: notifications.send"
      }
    },
    "/api/message-templates/:id": {
      "delete": {
        "tags": [
          "message-templates"
        ],
        "operationId": "message-templates_remove",
        "summary": "DELETE /api/message-templates/:id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: notifications.send"
      },
      "get": {
        "tags": [
          "message-templates"
        ],
        "operationId": "message-templates_findOne",
        "summary": "GET /api/message-templates/:id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: notifications.send"
      },
      "patch": {
        "tags": [
          "message-templates"
        ],
        "operationId": "message-templates_update",
        "summary": "PATCH /api/message-templates/:id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: notifications.send"
      }
    },
    "/api/message-templates/preview": {
      "post": {
        "tags": [
          "message-templates"
        ],
        "operationId": "message-templates_preview",
        "summary": "POST /api/message-templates/preview",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: notifications.send"
      }
    },
    "/api/message-templates/render/:key": {
      "post": {
        "tags": [
          "message-templates"
        ],
        "operationId": "message-templates_render",
        "summary": "POST /api/message-templates/render/:key",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: notifications.send"
      }
    },
...
    "/api/promo-codes": {
      "get": {
        "tags": [
          "promo-codes"
        ],
        "operationId": "promo-codes_findAll",
        "summary": "GET /api/promo-codes",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: promoCodes.manage"
      },
      "post": {
        "tags": [
          "promo-codes"
        ],
        "operationId": "promo-codes_create",
        "summary": "POST /api/promo-codes",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: promoCodes.manage"
      }
    },
    "/api/promo-codes/:id": {
      "delete": {
        "tags": [
          "promo-codes"
        ],
        "operationId": "promo-codes_deactivate",
        "summary": "DELETE /api/promo-codes/:id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: promoCodes.manage"
      },
      "get": {
        "tags": [
          "promo-codes"
        ],
        "operationId": "promo-codes_findOne",
        "summary": "GET /api/promo-codes/:id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: promoCodes.manage"
      },
      "patch": {
        "tags": [
          "promo-codes"
        ],
        "operationId": "promo-codes_update",
        "summary": "PATCH /api/promo-codes/:id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: promoCodes.manage"
      }
    },
    "/api/promo-codes/redeem": {
      "post": {
        "tags": [
          "promo-codes"
        ],
        "operationId": "promo-codes_redeem",
        "summary": "POST /api/promo-codes/redeem",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/referrals": {
      "get": {
        "tags": [
          "referral"
        ],
        "operationId": "referral_findAll",
        "summary": "GET /api/referrals",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: referrals.manage"
      }
    },
    "/api/referrals/:refereeId/qualify": {
      "post": {
        "tags": [
          "referral"
        ],
        "operationId": "referral_qualify",
        "summary": "POST /api/referrals/:refereeId/qualify",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: referrals.manage"
      }
    },
    "/api/referrals/apply": {
      "post": {
        "tags": [
          "referral"
        ],
        "operationId": "referral_apply",
        "summary": "POST /api/referrals/apply",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/referrals/mine": {
      "get": {
        "tags": [
          "referral"
        ],
        "operationId": "referral_mine",
        "summary": "GET /api/referrals/mine",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/referrals/my-code": {
      "get": {
        "tags": [
          "referral"
        ],
        "operationId": "referral_myCode",
        "summary": "GET /api/referrals/my-code",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/trips/:id": {
      "get": {
        "tags": [
          "trips"
        ],
        "operationId": "trips_findOne",
        "summary": "GET /api/trips/:id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: trips.read, trips.manage"
      }
    },
...
    "/api/trips/scheduled": {
      "get": {
        "tags": [
          "scheduled-trips"
        ],
        "operationId": "scheduled-trips_listAll",
        "summary": "GET /api/trips/scheduled",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "scheduled-trips"
        ],
        "operationId": "scheduled-trips_create",
        "summary": "POST /api/trips/scheduled",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/trips/scheduled/:id": {
      "delete": {
        "tags": [
          "scheduled-trips"
        ],
        "operationId": "scheduled-trips_cancel",
        "summary": "DELETE /api/trips/scheduled/:id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/trips/scheduled/mine": {
      "get": {
        "tags": [
          "scheduled-trips"
        ],
        "operationId": "scheduled-trips_listMine",
        "summary": "GET /api/trips/scheduled/mine",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}